etl/requirements.txt
fastapi
uvicorn[standard]
requests
tenacity
pandas
psycopg[binary]
SQLAlchemy
python-dotenv

httpx==0.27.0

psycopg2-binary==2.9.9
